Refactor file edit handling by updating imports and enhancing functionality ✏️
