+++ packages/core/src/genaisrc/genaiscript.d.ts
@@ -2045,6 +2045,11 @@ interface PromptGeneratorOptions extends ModelOptions, PromptSystemOptions {
      * Label for trace
      */
     label?: string
+
+    /**
+     * Write file edits to the file system
+     */
+    applyEdits?: boolean
 }
 
 interface FileOutputOptions {
